+++ source/string.c
@@ -75,6 +75,18 @@ int	ft_strcmp(const char *s1, const char *s2)
 	return (*(const unsigned char *)s1 - *(const unsigned char *)s2);
 }
 
+size_t	ft_strlen(char *s)
+{
+	size_t	i;
+
+	i = 0;
+	if (!s)
+		return (0);
+	while (s[i] != '\0')
+		i++;
+	return (i);
+}
+
 unsigned count_args(char   **args)
 {
     unsigned i;
